refactor: consolidate server fs helpers into fs.server
